Day 1 with pundit

+++ app/controllers/boards_controller.rb
@@ -1,6 +1,6 @@
 class BoardsController < ApplicationController
   def index
-    @boards = Board.includes(:lists)
+    @boards = policy_scope(Board).includes(:lists)
   end
 
   def new
@@ -9,6 +9,7 @@ class BoardsController < ApplicationController
 
   def create
     @board = Board.new(board_params)
+    @board.owner = current_user
     if @board.save
       redirect_to root_path, notice: 'Board was successfully created.'
     else
@@ -17,10 +18,12 @@ class BoardsController < ApplicationController
   end
 
   def edit
+    authorize board
     board
   end
 
   def update
+    authorize board
     if board.update(board_params)
       redirect_to root_path, notice: 'Board was successfully updated.'
     else
@@ -29,6 +32,7 @@ class BoardsController < ApplicationController
   end
 
   def destroy
+    authorize board
     board.destroy
     redirect_to root_path, notice: 'Board was successfully destroyed.'
   end
@@ -40,6 +44,6 @@ class BoardsController < ApplicationController
   end
 
   def board_params
-    params.require(:board).permit(:title)
+    params.require(:board).permit(:title, :owner_id)
   end
 end
